Move fixtures to conftest.py, test aardvark
